requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile - -o requirements.txt
aiohttp==3.9.4
    # via discord-py-self
aiosignal==1.3.1
    # via aiohttp
attrs==23.2.0
    # via aiohttp
discord-protos==0.0.2
    # via discord-py-self
discord-py-self==2.0.0
frozenlist==1.4.1
    # via
    #   aiohttp
    #   aiosignal
idna==3.7
    # via yarl
multidict==6.0.5
    # via
    #   aiohttp
    #   yarl
protobuf==5.26.1
    # via discord-protos
python-dotenv==1.0.1
yarl==1.9.4
    # via aiohttp
